Rework the LocalStorage implementation
